# docs/api/swagger.yaml
basePath: /v1
definitions:
...
  response.Envelope:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      success:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: moistello.com
info:
  contact:
    email: support@moistello.com
    name: Moistello Support
    url: https://moistello.com/support
  description: Decentralized savings circles on Stellar. REST API for circles, contributions,
    payouts, reputation, and governance.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://moistello.com/terms
  title: Moistello API
  version: 1.0.0
paths:
  /admin/audit-log:
    get:
      description: Returns the system audit log. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    entries:
                      type: array
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: '[Admin] Get audit log'
      tags:
      - Admin
  /admin/circles:
    get:
      description: Lists all circles with pagination, search, and status filter. Admin
        only.
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    circles:
                      type: array
                  type: object
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: '[Admin] List all circles'
      tags:
      - Admin
  /admin/feature-flags:
    post:
      consumes:
      - application/json
      description: Enables or disables a feature flag. Admin only.
      parameters:
      - description: Feature flag name and value
        in: body
        name: body
        required: true
        schema:
          properties:
            flag:
              type: string
            value:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    flag:
                      type: string
                    value:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: '[Admin] Update feature flag'
      tags:
      - Admin
...
  /admin/users:
    get:
      description: Lists all users with pagination and search. Admin only.
      parameters:
      - description: Search by wallet or email
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    users:
                      type: array
                  type: object
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: '[Admin] List users'
      tags:
      - Admin
  /auth/logout:
    post:
      description: Invalidates the current session.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Authentication
  /auth/me:
    post:
      description: Returns the authenticated user's profile. Requires Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    user:
                      type: object
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Authentication
  /auth/nonce:
    post:
      consumes:
      - application/json
      description: Returns a signed nonce for wallet authentication. The nonce must
        be signed with the wallet's private key and sent to /auth/verify.
      parameters:
      - description: Wallet address
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    nonce:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Get authentication nonce
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access token and refresh
        token pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    refreshToken:
                      type: string
                    token:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Refresh JWT tokens
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Verifies wallet signature and creates a user account with optional
        profile fields (displayName, email, countryCode, language). Returns JWT tokens.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    refreshToken:
                      type: string
                    token:
                      type: string
                    user:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Register new user with profile
      tags:
      - Authentication
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verifies a signed nonce to prove wallet ownership. Creates a user
        account if one doesn't exist. Returns JWT tokens.
      parameters:
      - description: Signature payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    refreshToken:
                      type: string
                    token:
                      type: string
                    user:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Verify wallet signature and login
      tags:
      - Authentication
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /auth/verify or /auth/register. Format:
      "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"